+++ src/output/help.rs
@@ -969,7 +969,12 @@ pub(crate) fn dimensions() -> (Option<usize>, Option<usize>) {
     #[cfg(feature = "wrap_help")]
     terminal_size::terminal_size()
         .map(|(w, h)| (Some(w.0.into()), Some(h.0.into())))
-        .unwrap_or((None, None))
+        .unwrap_or_else(|| (parse_env("COLUMNS"), parse_env("LINES")))
+}
+
+#[cfg(feature = "wrap_help")]
+fn parse_env(var: &str) -> Option<usize> {
+    std::env::var_os(var)?.to_str()?.parse::<usize>().ok()
 }
 
 fn should_show_arg(use_long: bool, arg: &Arg) -> bool {